refactor: 🎨 more precise comment more idomatic rust
