+++ crates/mako/src/transform.rs
@@ -29,7 +29,7 @@ use crate::transformers::transform_px2rem::Px2Rem;
 use crate::transformers::transform_react::mako_react;
 use crate::transformers::transform_try_resolve::TryResolve;
 use crate::transformers::transform_virtual_css_modules::VirtualCSSModules;
-use crate::visitors::named_export_default::NamedExportDefault;
+use crate::visitors::default_export_namer::DefaultExportNamer;
 
 pub struct Transform {}
 
@@ -65,7 +65,7 @@ impl Transform {
                         file.extname == "tsx" || file.extname == "jsx" || file.extname == "js";
                     if context.args.watch && !file.is_under_node_modules && could_be_react_component
                     {
-                        visitors.push(Box::new(NamedExportDefault::new()));
+                        visitors.push(Box::new(DefaultExportNamer::new()));
                     }
                     // TODO: refact mako_react
                     visitors.push(Box::new(mako_react(
