+++ crates/mako/src/visitors/mod.rs
@@ -1,3 +1,3 @@
 pub(crate) mod css_dep_analyzer;
+pub(crate) mod default_export_namer;
 pub(crate) mod js_dep_analyzer;
-pub(crate) mod named_export_default;
